requirements.txt
openai
Flask
python-dotenv
Flask-CORS
requests
pymupdf